Modifications after GRAMPS devs comments
- The pages style is now based on the bootstrap framework
This framwork is widely used over the internet, and IMHO it is a good
thing to benefit from other people's work.
With bootstrap, a big part of the CSS job is already done.
Moreover, with bootstrap, it is easy to build pages with a responsive
design: The page aspect is adapted to the screen size. Could be nice for
smartphones.
The drawback is that the rendering is not identical to the narrative web
report.
So, for fans of Mainz style (yellowish old parchment): It tried to make
a version that looks like it, but it is not strictly the same.

- Sources, Media, Places and Repositories pages: index table was wider
than page header.
I included the "DataTables" Javascript library, which includes the
responsive datatables feature.
Responsive data tables dynamically hide/show columns depending on the
page width.
Unfortunately, the library has a bug and it doesn't really work yet, but
I am confident that it will be solved in near future.

- thumbnails are added in the media index.